refactor(voice/capture): extract restart-verdict types into capture subpackage [T1.4 step 1]
Master mission Phase 1 / T1.4 — first stage of the
``_capture_task.py`` (2785 LOC) split. The spec marks the full
split as a "DEDICATED SESSION" because of the mixin surgery +
threading invariants + 13 timing-primitive test patches; this
commit takes the lowest-risk first slice so subsequent commits
can build on the established subpackage structure without
attempting the entire split in one go.

What lands:

  * New subpackage ``src/sovyx/voice/capture/`` with
    ``__init__.py`` re-exporting the public surface.
  * New module ``capture/_restart.py`` containing the four
    restart-verdict ``StrEnum``s (Exclusive/Shared on Windows,
    AlsaHwDirect/SessionManager on Linux), the four matching
    ``@dataclass(frozen=True, slots=True)`` result types, two
    Linux host-API constants, and the four metric-emitter
    functions. ~380 LOC total — pure data + pure functions, no
    class state coupling, no awaits, no I/O on the hot path.
  * ``_capture_task.py`` (2785 → ~2400 LOC) drops the extracted
    definitions and re-exports every name via the explicit
    ``import X as X`` pattern (mypy strict requires this form
    for an import to count as a public re-export). This
    preserves every existing import like
    ``from sovyx.voice._capture_task import ExclusiveRestartVerdict``
    AND the 13 timing-primitive test patches that target
    ``sovyx.voice._capture_task.<name>`` (CLAUDE.md anti-pattern
    #20).
  * ``import dataclass`` and ``import StrEnum`` removed from
    ``_capture_task.py`` (no longer used after the move).

What does NOT land (deferred to subsequent commits):

  * The ``AudioCaptureTask`` class itself (~1972 LOC, the bulk
    of the file). Splitting via the ``_RestartMixin`` pattern
    is the higher-risk surgery the spec calls out; doing it as
    a separate commit lets reviewers focus on one thing at a
    time and lets a future session pick up cleanly if context
    runs out mid-mixin-extraction.
  * The remaining sections — sustained-underrun detector
    constants, session-manager contention helpers, exception
    classes, ``_rms_db_int16``, ``_PEAK_DB_RE`` /
    ``_extract_peak_db``, ``_resolve_input_entry``. Each is a
    candidate for a future extraction (``_loop.py``, ``_ring.py``,
    ``_epoch.py`` per spec target).

Behavioural delta: zero. The extracted symbols are referenced by
their original names through the re-export; runtime semantics
are unchanged. mypy strict (420 source files), ruff, ruff format,
bandit, and the full pytest suite all pass with zero test
modifications.

Validation:
  * ruff check + format clean
  * mypy strict: 0 issues across 420 source files (added the
    new ``voice.capture`` subpackage, +2 source files)
  * bandit clean
  * pytest full suite: 11894 passed, 25 skipped, 57 deselected
    (zero delta — pure structural extraction; no test changes
    needed because the legacy import surface is preserved
    verbatim via ``import X as X`` re-exports)

T1.4 progress: step 1 of N. Subsequent commits land additional
extractions until the ``AudioCaptureTask`` class itself is
split via the ``_RestartMixin`` pattern. Each step is gates-
green and leaves the codebase coherent so a fresh session can
pick up from any landed step.

Mission: docs-internal/missions/MISSION-voice-final-skype-grade-2026.md §Phase 1.T1.4 (step 1)
